can switch to contact form in case of contact without email
