+++ vector/src/main/java/im/vector/fragments/PeopleFragment.java
@@ -21,6 +21,7 @@ import android.content.ContentResolver;
 import android.content.DialogInterface;
 import android.content.pm.PackageManager;
 import android.content.Intent;
+import android.content.Context;
 import android.database.Cursor;
 import android.net.Uri;
 import android.os.AsyncTask;
@@ -39,6 +40,7 @@ import android.view.ViewGroup;
 import android.widget.CheckBox;
 import android.widget.Filter;
 import android.provider.ContactsContract;
+import android.app.Activity;
 
 import org.matrix.androidsdk.MXDataHandler;
 import org.matrix.androidsdk.MXSession;
@@ -508,6 +510,7 @@ public class PeopleFragment extends AbsHomeFragment implements ContactsManager.C
          }
         else {// tell the user that the email must be filled. Propose to fill it
             if (ContactsManager.getInstance().isContactBookAccessAllowed()) {
+                final Context theContext = this.getContext();
                 //enterEmailAddress(item.mContact);
                 AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(getActivity());
                 alertDialogBuilder.setMessage(getString(R.string.people_invalid_warning_msg));
@@ -518,7 +521,7 @@ public class PeopleFragment extends AbsHomeFragment implements ContactsManager.C
                         .setPositiveButton(R.string.action_edit_contact_form,
                                 new DialogInterface.OnClickListener() {
                                     public void onClick(DialogInterface dialog, int id) {
-                                        editContactForm(item.mContact);
+                                        DinsicUtils.editContactForm(theContext,getActivity(),getString(R.string.people_edit_contact_warning_msg),item.mContact);
                                     }
                                 });
 
@@ -553,72 +556,6 @@ public class PeopleFragment extends AbsHomeFragment implements ContactsManager.C
 
     }
 
-    /**
-     * Edit contact form
-     */
-    private void editContactForm(final Contact contact ) {
-        LayoutInflater inflater = LayoutInflater.from(this.getContext());
-
-        Cursor namesCur=null;
-        boolean switchToContact=false;
-        try
-        {
-            ContentResolver cr = getContext().getContentResolver();
-            namesCur = cr.query(ContactsContract.Data.CONTENT_URI,
-                    new String[]{ContactsContract.Contacts.DISPLAY_NAME_PRIMARY,
-                            ContactsContract.Contacts.LOOKUP_KEY,
-                            ContactsContract.CommonDataKinds.StructuredName.CONTACT_ID
-                    },
-                    ContactsContract.Data.MIMETYPE + " = ? AND "+ ContactsContract.Data.CONTACT_ID + " = ?",
-
-                    new String[]{ContactsContract.CommonDataKinds.StructuredName.CONTENT_ITEM_TYPE,
-                            contact.getContactId()}, null);
-            if (namesCur != null) {
-                if (namesCur.moveToNext()) {
-                    String contactId = namesCur.getString(namesCur.getColumnIndex(ContactsContract.CommonDataKinds.StructuredName.CONTACT_ID));
-                    Uri mSelectedContactUri;
-                    int mLookupKeyIndex;
-                    int mIdIndex;
-                    String mCurrentLookupKey;
-                    mLookupKeyIndex = namesCur.getColumnIndex(ContactsContract.Contacts.LOOKUP_KEY);
-                    mCurrentLookupKey = namesCur.getString(mLookupKeyIndex);
-                    //mCursor.getColumnIndex(ContactsContract.Contacts._ID);
-                    long mCurrentId = Integer.parseInt(contactId);//namesCur.getLong(mIdIndex);
-                    mSelectedContactUri =
-                            ContactsContract.Contacts.getLookupUri(mCurrentId, mCurrentLookupKey);
-
-
-                    Intent editIntent = new Intent(Intent.ACTION_EDIT);
-                    editIntent.setDataAndType(mSelectedContactUri, ContactsContract.Contacts.CONTENT_ITEM_TYPE);
-                    editIntent.putExtra("finishActivityOnSaveCompleted", true);
-                    startActivity(editIntent);
-                    switchToContact=true;
-                }
-            }
-        } catch (Exception e) {
-            Log.e(LOG_TAG, "## editContactForm(): Exception - Msg=" + e.getMessage());
-        }
-        if (!switchToContact){
-            AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(getActivity());
-            alertDialogBuilder.setMessage(getString(R.string.people_edit_contact_warning_msg));
-
-            // set dialog message
-            alertDialogBuilder
-                    .setCancelable(false)
-                    .setPositiveButton(R.string.ok,
-                            new DialogInterface.OnClickListener() {
-                                public void onClick(DialogInterface dialog, int id) {
-
-                                }
-                            });
-
-            // create alert dialog
-            AlertDialog alertDialog = alertDialogBuilder.create();
-            // show it
-            alertDialog.show();
-
-        }
-    }
